moved restart_unit from control_manager.py to service_manager
