+++ systemctl_mqtt/_dbus/service_manager.py
@@ -16,9 +16,10 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import jeepney
-
+import logging
 import systemctl_mqtt._dbus
 
+_LOGGER = logging.getLogger(__name__)
 
 class ServiceManager(jeepney.MessageGenerator):
     """
@@ -41,6 +42,10 @@ class ServiceManager(jeepney.MessageGenerator):
             remote_obj=self, method="GetUnit", signature="s", body=(name,)
         )
 
+    def RestartUnit(self, name: str, mode: str) -> jeepney.low_level.Message:
+        return jeepney.new_method_call(
+            remote_obj=self, method="RestartUnit", signature="ss", body=(name,mode,)
+        )
 
 class Unit(systemctl_mqtt._dbus.Properties):  # pylint: disable=protected-access
     """
@@ -55,3 +60,24 @@ class Unit(systemctl_mqtt._dbus.Properties):  # pylint: disable=protected-access
         super().__init__(object_path=object_path, bus_name="org.freedesktop.systemd1")
 
     # pylint: disable=invalid-name
+
+def restart_unit(unit_name: str):
+    proxy = get_systemd_manager_proxy()
+    try:
+        reply = proxy.RestartUnit(unit_name, "replace")
+        _LOGGER.info(f"called successfully: {reply}")
+    except Exception as e:
+        _LOGGER.error(f"Failed to call: {e}")
+
+def get_systemd_manager_proxy() -> jeepney.io.blocking.Proxy:
+    # https://jeepney.readthedocs.io/en/latest/integrate.html
+    # https://gitlab.com/takluyver/jeepney/-/blob/master/examples/aio_notify.py
+    return jeepney.io.blocking.Proxy(
+        msggen=ServiceManager(),
+        connection=jeepney.io.blocking.open_dbus_connection(
+            bus="SYSTEM",
+            # > dbus-broker[…]: Peer :1.… is being disconnected as it does not
+            # . support receiving file descriptors it requested.
+            enable_fds=True,
+        ),
+    )
